tests/prctl.sh: match strace output with -a value more than 79
For some tests it is instructive to set -a to value more than 79; the
issue is that this alignment width is more than the marker output, which
leads to a match failure.  To avoid that, allow for arbitrary number
of spaces before the syscall return value.

* tests/prctl.sh: Relax the regular expression to allow for arbitrary
(non-zero) number of spaces.